move SVG to a single file
