+++ src/frontend/components/shared/SVG.tsx
@@ -1,0 +1,12 @@
+// import EditIcon from '../../../../../public/img/foodtrucks/edit.svg'
+import EditSVG from '../../../../public/img/icons/edit.svg'
+import CommentSVG from '../../../../public/img/icons/message-circle.svg'
+import StarSVG from '../../../../public/img/icons/star.svg'
+import MapSVG from '../../../../public/img/icons/map.svg'
+import InfoSVG from '../../../../public/img/icons/info.svg'
+
+export const EditIcon = EditSVG
+export const CommentIcon = CommentSVG
+export const StarIcon = StarSVG
+export const MapIcon = MapSVG
+export const InfoIcon = InfoSVG
